Added a data changed flag, make fields private, added file selector
A data changed flag was added so that all the other methods could set
this to signal that the data had been changed. This way the program can
prompt for confirmation before quitting or loading another data file.

Made most of the fields private and non-static. Also added a file
selector for file loading